requirements.txt
aiohttp
dnspython
motor
psutil
pyrogram
python-dotenv
requests
tgcrypto
